Fix or silence (msvc) compiler warnings about narrowing conversion, unused variables and formal parameters, , name hiding types and othe rminor stuff
(cherry picked from commit ca4e2c0e4d7977364099092e63aad674e4b4bf97)

+++ test/maybe.cpp
@@ -14,7 +14,7 @@ BOOST_PARAMETER_KEYWORD(tag, unused)
 template <class Args>
 int f(Args const& args)
 {
-    return args[kw | 1.f];
+    return static_cast<int>(args[kw | 1.f]);
 }
 
 } // namespace test
